+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ### Changes
-- Buff rennet drops
-- Add config for classic sprinkler behavior
-- Disallow knapping with rotten pumpkins
+- Made behavior of uninsulated top ovens more consistent. Now they will heat items, but only to half the requested temperature.
+- Added a note about the max greenhouse size to the book.
+- Replaced uses of the Empty Jar With Lid item with a tag internally.